nixtla/timegpt/requirements.txt
pandas
numpy
nixtla
uvicorn
fastapi
pydantic
ouro-py
python-dotenv
autots
tabulate